refactor(search-engine): Extract core search types to shared package
This commit migrates core search-related type definitions (such as `ISearchProvider`, `TuffSearchResult`, `ITuffGatherOptions`, etc.) from the `search-engine` module within `apps/core-app` to the shared `@talex-touch/utils` package.

The primary goals of this refactoring are:
- To improve code modularity and reusability.
- To decouple the core search logic from the main application, laying the groundwork for future extensions.
- To standardize type definitions across the entire project, making them clearer and easier to maintain.

Associated changes include:
- Re-enabling `fileProvider` in `search-core`.
- Cleaning up debug log outputs in the `storage` module.

BREAKING CHANGE: Core search types (e.g., `ISearchProvider`) have moved from `apps/core-app` to the `@talex-touch/utils` package. All internal modules depending on these types will need to update their import paths.
