+++ apps/core-app/src/main/modules/storage/index.ts
@@ -65,7 +65,6 @@ export class StorageModule extends BaseModule {
     }
 
     const p = path.resolve(this.filePath!, name)
-    console.log(`[Config] Get config ${name} from ${p}`)
     const file = fse.existsSync(p) ? JSON.parse(fse.readFileSync(p, 'utf-8')) : {}
 
     this.configs.set(name, file)
@@ -164,7 +163,6 @@ export class StorageModule extends BaseModule {
 
     channel.regChannel(ChannelType.MAIN, 'storage:get', ({ data }) => {
       if (!data || typeof data !== 'string') return {}
-      console.log(`[Config] Get config ${data}`)
       return this.getConfig(data)
     })
