+++ packages/utils/common/index.ts
@@ -1,2 +1,3 @@
 export * from './storage/index'
 export * from './utils'
+export * from './search'
