+++ packages/utils/core-box/tuff/tuff-dsl.ts
@@ -16,6 +16,8 @@
  * @module core-box/tuff-dsl
  */
 
+import { TalexTouch } from "packages/utils/types";
+
 /**
  * 定义高亮范围
  * @description 右开区间 [start, end)
@@ -1108,6 +1110,16 @@ export interface TuffQuery {
 }
 
 /**
+ * Represents the statistics for a single sort middleware.
+ */
+export interface SortStat {
+  /** The name of the sorting middleware. */
+  name: string
+  /** The time taken by the middleware in milliseconds. */
+  duration: number
+}
+
+/**
  * 搜索结果结构
  *
  * @description
@@ -1170,6 +1182,9 @@ export interface TuffSearchResult {
    * The provider(s) to activate after this search result.
    */
   activate?: IProviderActivate[];
+
+  /** Optional statistics about the sorting process. */
+  sort_stats?: SortStat[]
 }
 
 export interface IProviderActivate {
@@ -1178,6 +1193,95 @@ export interface IProviderActivate {
   icon?: TuffIcon
   time?: number
   meta?: Record<string, any>
+}
+
+
+/**
+ * Defines the interface for a sort middleware.
+ * Each middleware receives an array of items and should return a sorted array.
+ */
+export interface ISortMiddleware {
+  /** A unique name for the middleware, used for logging and stats. */
+  readonly name: string
+  /**
+   * The sort function that processes the items.
+   * @param items - The array of TuffItems to be sorted.
+   * @param query - The original search query for context.
+   * @param signal - An AbortSignal to cancel the sorting operation.
+   * @returns A sorted array of TuffItems.
+   */
+  sort(items: TuffItem[], query: TuffQuery, signal: AbortSignal): TuffItem[]
+}
+
+/**
+ * Search Provider Interface (formerly ISearchSource)
+ *
+ * Defines the contract for any module that provides search results to the engine.
+ * It's a simplified, stateless interface focused solely on providing results for a given query.
+ */
+export interface IExecuteArgs {
+  item: TuffItem
+  searchResult?: TuffSearchResult
+}
+
+export interface ISearchProvider<C> {
+  /**
+   * Unique identifier for the provider, e.g., "mac-applications", "file-system", "clipboard-history"
+   * @required
+   */
+  readonly id: string
+
+  /**
+   * The type of the source, used for categorization and filtering.
+   * @required
+   */
+  readonly type: TuffSourceType
+
+  /**
+   * User-friendly name for the provider, displayed in settings or logs.
+   */
+  readonly name?: string
+
+  /**
+   * Icon for the provider.
+   */
+  readonly icon?: any
+
+  /**
+   * Core search method (PULL mode).
+   * The engine calls this method to get results from the provider.
+   *
+   * @param query - The search query object, containing text and other context.
+   * @param signal - An AbortSignal to cancel the search operation.
+   * @returns A promise that resolves to a full TuffSearchResult object, allowing the provider
+   * to influence the final result, including the next activation state.
+   */
+  onSearch(query: TuffQuery, signal: AbortSignal): Promise<TuffSearchResult>
+
+  /**
+   * Optional method to handle activation.
+   * Called when the provider is prioritized, e.g., via an activation keyword.
+   */
+  onActivate?(): void
+
+  /**
+   * Optional method to handle deactivation.
+   */
+  onDeactivate?(): void
+
+  /**
+   * Optional method to execute an item.
+   * @param args The arguments for execution, including the item and search context.
+   * @returns A promise that resolves to an activation object if the provider should be
+   * activated, or `null` otherwise.
+   */
+  onExecute?(args: IExecuteArgs): Promise<IProviderActivate | null>
+
+  /**
+   * Optional method to load provider.
+   * @param context The context of the provider.
+   */
+  onLoad?(context: C): Promise<void>
 }
 
 // ==================== 插件接口预览 ====================
